Delete Dockerfile
I see no material change to the Dockerfile other than changing the repo to your Github repo. You are free to fork the project and then publish the Dockerfile there. But to keep things in order, the Dockerfile in this repository should be pointing to the code in this repo.